\name{getAnaClass}
\alias{getAnaClass}
\title{ generic list for analysis input on one KIND of variable}
\description{ Generation of a generic list with the needed input
  parameters to specify how the assimilation analysis is to be conducted
  on a specific variable KIND
}
\usage{
getAnaClass()
}


