\name{getCostTheta}
\alias{getCostTheta}
\title{ Get a cost background term}
\description{ From a theta LISt containing the background information of
  the control varaibles, the function obtains the cost background term
}
\usage{
getCostTheta(theta)
}
\arguments{
  \item{theta}{LIST, as defined in the package rDAF}
}
\value{
A scalar
}
\author{
Javier Garcia-Pintado
}
