\name{getOV}
\alias{getOV}
\title{ Get a list with overflow information}
\description{
  Get a structure with overflow parameterizations in physical space. These are extracted from the Technical report by Danabasoglu et al. (2010).
  TODO: only Denmark strait is included here. Remaining overflows have to be included.
}
\usage{
getOV()
}
\value{
  A list
}
