\name{plotOV}
\alias{plotOV}
\title{ Add overflow metadata to an existing plot}
\description{
 Add rectangles indicating source, entrainment, and interior overflow
 regions to an existing plot
}
\usage{
plotOV(OV)
}
\arguments{
  \item{OV}{An overflow list as produced by \code{getOV}}
}
\details{
See CESM documentation about overflows
}
\value{
  Symbols added to a plot
}
