\name{POSIXt2BPstr}
\alias{POSIXt2BPstr}
\title{ Months between a POSIXct time to Before Present}
\description{
 Before Presetn times is defined as the start of 1950. The function
 returns the ellapsed months between a POSIXct timestamp and BP time.
}
\usage{
POSIXt2BPstr(x)
}
\arguments{
  \item{x}{POSIXt [Anno Domini]}
}
