\name{rdafCESM-package}
\alias{rdafCESM-package}
\alias{rdafCESM}
\docType{package}
\title{
\packageTitle{rdafCESM}
}
\description{
\packageDescription{rdafCESM}
}
\details{
The key function in this package in cesmDA(), which refers to a
example of use provided in the package. For detailed analysis of CESM
output, the user is referred to NCL and python utilities developed by NCAR. This R package is by far more
limited with respect to CESM output analysis, but it helps to organise CESM input and
evaluate prototype assimilation schemes for climatic data CESM within R.
}
\author{
\packageAuthor{rdafCESM}

Maintainer: \packageMaintainer{rdafCESM}
}
\references{
Garcia-Pintado and Paul (2018), Evaluation of iterative Kalman smoother
schemes for multi-decadal past climate analysis with comprehensive earth
system models, Geoscientific Model development.
}
\keyword{ package }
\keyword{ data assimilation }
\keyword{ climate }
\seealso{
 \code{\link{rDAF}}
}
