\name{todPOSIXct}
\alias{todPOSIXct}
\title{ Get START_TOD variable in CESM run from a POSIXct time}
\description{
 envrun$START_TO indicates the hour of the run start time in 
}
\usage{
todPOSIXct(x)
}
\arguments{
  \item{x}{POSIXct}
}
\details{
See CESM documentation about this variable
}
\value{
  A CHARACTER variable
}
